🚧 WIP: additional refactor
- add route folders
- update `app.py`
- update `rollbar.py`
