+++ backend-flask/lib/rollbar.py
@@ -1,11 +1,10 @@
-from flask import current_app as app
 from flask import got_request_exception
 from time import strftime
 import os
 import rollbar
 import rollbar.contrib.flask
 
-def init_rollbar():
+def init_rollbar(app):
   rollbar_access_token = os.getenv('ROLLBAR_ACCESS_TOKEN')
   rollbar.init(
       # access token
